fcdbus: raise an exception in case the user's home is not created
The current code tries to create the user's home directory in case its
not yet created. However, as the fcdbus is ran as the user logged into
cockpit and not as root, trying to create the home directory is
something that will always fail.

Instead of trying to solve this misconfiguration issue by ourselves,
let's inform the user the paths they could take in order to properly fix
the issue.

Resolves:
https://github.com/fleet-commander/fc-admin/issues/192

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>