Add Dialog classes
As for now, there are several basic methods for operating common
dialogs. These methods include only the display of such messages.
For the other actions, the caller of dialog has to do it in his
own way. For example, there is no central point for closing or
event handling.

This patch adds base dialog class and several derived. Now we can
easily manage spinner, question and message dialogs in one place.

Fixes: https://github.com/fleet-commander/fc-admin/issues/222
Signed-off-by: Stanislav Levin <slev@altlinux.org>