tests: Make use of sorting of XML attributes
> Prior to Python 3.8, the serialisation order of the XML attributes of
elements was artificially made predictable by sorting the attributes by
their name. Based on the now guaranteed ordering of dicts, this
arbitrary reordering was removed in Python 3.8 to preserve the order in
which attributes were originally parsed or created by user code.

As of Python3.8:
> Canonicalization is a way to normalise XML output in a way that allows
byte-by-byte comparisons and digital signatures. It reduced the freedom
that XML serializers have and instead generates a more constrained XML
representation. The main restrictions regard the placement of namespace
declarations, the ordering of attributes, and ignorable whitespace.

In order to support Python < 3.8 the custom sorting is applied.

Fixes: https://github.com/fleet-commander/fc-admin/issues/248
Signed-off-by: Stanislav Levin <slev@altlinux.org>