fcadmin: Improve Python logging
The logging shouldn't be used before configuring handlers.
Otherwise, the logger functions such debug(), info(), warning(), error()
and critical() will call basicConfig() automatically if no handlers are
defined for the root logger [0].

This happened on importing of admin.fleetcommander.fcad and resulted in
the logging configuration was skipped(level and formatting).

According to [1]:
    A good convention to use when naming loggers is to use a module-level
    logger, in each module which uses logging, named as follows:

    logger = logging.getLogger(__name__)

    This means that logger names track the package/module hierarchy, and
    it’s intuitively obvious where events are logged just from the logger
    name.

Note. For scripts or modules not intended for import the logger's name
will be '__main__'. This is not convenient and for these cases the
logger's name will be set as module's filename:
```
logger = logging.getLogger(os.path.basename(__file__))
```

[0]: https://docs.python.org/3/library/logging.html#logging.basicConfig
[1]: https://docs.python.org/3/howto/logging.html#advanced-logging-tutorial

Signed-off-by: Stanislav Levin <slev@altlinux.org>